Add rounds 4-5: 24h dip-in-trend strategy, holdout-coin test, sensitivity, DEX notes
Round 4 sweeps holding period for the trend family and a 12-cell grid for
dip-in-trend; round 5 tests the fixed specs on 12 coins never used in
selection. The trend-filtered 24h dip held 24h is positive in 8 of 9 years
on both panels; its long-only form is positive on all 12 holdout coins.
README updated with the final spec, evidence tables and SushiSwap notes.

+++ research/crypto-strategy-screen/bt.py
@@ -14,7 +14,7 @@ def load_hourly(path):
     df = pd.DataFrame(rows, columns=["t","low","high","open","close","volume"])
     df["t"] = pd.to_datetime(df["t"], unit="s", utc=True)
     df = df.set_index("t").sort_index(); df = df[~df.index.duplicated()]
-    return df
+    return df.astype(float)
 
 def load_coin(c):
     import os
